[python] Split monolithic python_dict_handler.cpp into modules
Break the 3715-line python_dict_handler.cpp into eight focused
translation units under src/python-frontend/python-dict/ (core, nested,
construction, access, type resolution, assignment, methods, compare)
sharing one python_dict_internal.h include bundle. Pure mechanical
refactor: every method body is moved verbatim and each file-local
helper stays in the translation unit that calls it. Behaviour is
unchanged; all 250 Python dict regression tests pass.
